Adding backtrace to MismatchError
`thiserror` exposes backtrace via `std::backtrace::Backtrace` which is
currently an experimental API. In order to use it we need to enable
nightly which is set in `rust-toolchain`.

We originally had `VictorsError` derive `PartialEq` however binary
operation `==` cannot be applied to `Backtrace`. We still want
`VectorsErrors` to implement `PartialEq` to make comparisons easy such
as using `assert_eq` in tests so we manually impement it.