Feature Parity Work with Scientist
This commit is focused around furthering parity with the Ruby
Scientist gem. The primary areas of focus are the following

* Custom ignore fn - allow users to specify rules to ignore mismatches
* `err_on_mismatches` flag on `Experiment` to return an error on mismatches
* Custom value and error comparators to allow users to alter how.
`Observation` equivalency works. Some of this work is still in progress.
* Constructing `ExperimentResult` - Previously we were not providing.
mismatched or ignored observations. In order to avoid issues with Rust's.
borrow checker I ended up storing indexes for control (still looking for
a better name) as well as mismatces and ignores. I provide functions to
then get the underlying observation data from those indexes. Not sure if
there there is a way to not need to fall back to indexes or not. Will
probably do additional research.