Fix Test Issue With Captured Variable in Publisher Closure
Fixes issue getting experiment result from publisher via captured
closure. It was complaining about the captured variable not living long
enough. This was caused by the variable ordering. I had instantiated the
variable that gets captured after the lifetime associated with the boxed
closure. Moving the instatiation of the RefCell to before creating the
experiment solved the issue.

I should learn to pay more attention to the
error details:

"Note: values in a scope are dropped in the opposte order they are
defined"